Restructuring of folder
