+++ ts-tests/common/abi/bridge/GenericHandler.json
