+++ ts-tests/common/abi/precompile/staking.json
