+++ ts-tests/common/setup/register-parathread.ts
@@ -3,7 +3,7 @@ import '@polkadot/api-augment';
 import { ApiPromise, Keyring, WsProvider } from '@polkadot/api';
 import { TypeRegistry } from '@polkadot/types/create';
 import { Bytes } from '@polkadot/types';
-import { loadConfig, signAndSend } from './utils';
+import { loadConfig, signAndSend } from '../utils';
 
 async function registerParathread(api: ApiPromise, config: any) {
     // Get keyring of Alice, who is also the sudo in dev chain spec
