+++ ts-tests/common/setup/setup-enclave.ts
@@ -1,9 +1,6 @@
-import fs from 'fs';
 import '@polkadot/api-augment';
 import { ApiPromise, Keyring, WsProvider } from '@polkadot/api';
-import { TypeRegistry } from '@polkadot/types/create';
-import { Bytes } from '@polkadot/types';
-import { loadConfig, signAndSend } from './utils';
+import { loadConfig, signAndSend } from '../utils';
 import { hexToU8a } from '@polkadot/util';
 
 const mrenclave = process.argv[2];
@@ -14,7 +11,7 @@ async function setAliceAsAdmin(api: ApiPromise, config: any) {
     const keyring = new Keyring({ type: 'sr25519' });
     const alice = keyring.addFromUri('//Alice');
 
-    const tx = api.tx.sudo.sudo(api.tx.teerex.setAdmin('esqZdrqhgH8zy1wqYh1aLKoRyoRWLFbX9M62eKfaTAoK67pJ5')); 
+    const tx = api.tx.sudo.sudo(api.tx.teerex.setAdmin('esqZdrqhgH8zy1wqYh1aLKoRyoRWLFbX9M62eKfaTAoK67pJ5'));
 
     console.log(`Setting Alice as Admin for Teerex`);
     return signAndSend(tx, alice);
@@ -23,11 +20,11 @@ async function setAliceAsAdmin(api: ApiPromise, config: any) {
 async function updateScheduledEnclave(api: ApiPromise, config: any) {
     const keyring = new Keyring({ type: 'sr25519' });
     const alice = keyring.addFromUri('//Alice');
-    
+
     const tx = api.tx.teerex.updateScheduledEnclave(block, hexToU8a(`0x${mrenclave}`));
 
-    console.log("Schedule Enclave Extrinsic sent"); 
-    return signAndSend(tx, alice)
+    console.log('Schedule Enclave Extrinsic sent');
+    return signAndSend(tx, alice);
 }
 
 (async () => {
@@ -41,11 +38,10 @@ async function updateScheduledEnclave(api: ApiPromise, config: any) {
 
     await setAliceAsAdmin(api, config);
     await updateScheduledEnclave(api, config);
-    
+
     await api.disconnect();
     provider.on('disconnected', () => {
         console.log('Disconnect from relaychain');
         process.exit(0);
     });
 })();
-
