--- ts-tests/tests/global.d.ts
@@ -1,1 +1,0 @@
-declare module 'solc';
